chore: rename Go module from go.miloapis.com/bgp to go.miloapis.com/cosmos
Renames the Go module path, the cmd/bgp entrypoint directory to cmd/cosmos,
and updates the Dockerfile binary name and entrypoint to match.
