+++ internal/controller/provider_reconciler.go
@@ -17,11 +17,11 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
-	providersv1alpha1 "go.miloapis.com/bgp/api/providers/v1alpha1"
-	bgpv1alpha1 "go.miloapis.com/bgp/api/bgp/v1alpha1"
-	"go.miloapis.com/bgp/internal/provider"
-	frrprovider "go.miloapis.com/bgp/internal/provider/frr"
-	gobgpprovider "go.miloapis.com/bgp/internal/provider/gobgp"
+	providersv1alpha1 "go.miloapis.com/cosmos/api/providers/v1alpha1"
+	bgpv1alpha1 "go.miloapis.com/cosmos/api/bgp/v1alpha1"
+	"go.miloapis.com/cosmos/internal/provider"
+	frrprovider "go.miloapis.com/cosmos/internal/provider/frr"
+	gobgpprovider "go.miloapis.com/cosmos/internal/provider/gobgp"
 )
 
 const (
